+++ src/main/java/kr/hhplus/be/server/interfaces/point/PointRequest.java
@@ -2,11 +2,13 @@ package kr.hhplus.be.server.interfaces.point;
 
 import kr.hhplus.be.server.application.point.PointCommand;
 
+import java.math.BigDecimal;
+
 public record PointRequest() {
 
     public record Charge(
             long userId,
-            long amount
+            BigDecimal amount
     ) {
     }
